Pallet crowdsale test coverage (#799)
* initialized TRN crowdsale pallet

* upd cargo.lock

* added asset_exists function to InspectExt trait

* added vec import to nft impls - fixed external dep compilation error

* updated redeem, added todos, updates structs

* Add mock + tests for initialize (#791)

* Add mock + tests for initialize

* Few tidy ups

* Add impls file, calculate_voucher_rewards, and on_initialize logic

* Update voucher_rewards calc to account for decimals partially (Needs some thinking)

* minor improvements + setup

* updated logic for burning vouchers

---------


* Pallet crowdsale upd (#793)

* added log target, removed decimals const, removed redundant enums

* fixed compilation

* fix div by zero bug

* fixed decimals burn calc logic

* updated close_sales_at with concrete example and todos

* cargo fmt --all

* Re introduce decimals, some testing work for crowdsale rounding

---------


* Crowdsale voucher rewards calculations (#794)

* added log target, removed decimals const, removed redundant enums

* fixed compilation

* fix div by zero bug

* fixed decimals burn calc logic

* updated close_sales_at with concrete example and todos

* cargo fmt --all

* Re introduce decimals, some testing work for crowdsale rounding

* Add some tests

* Add better calculation function

* Add comments and better tests

* Remove extra comment in assetsExt

* Change input type to U256 for contributions

* Added Result return type for reward calculation

---------


* Pallet crowdsale upd (#795)

* added sp io dep

* - updated SaleInformation to add vault account
- updated SaleStatus to also display block num with status variants

* - updated storage types
- updated events and errors
- added code for offchain worker
- added claim function
- refactor logic

* SaleEndBlocks back to option query, payout admin account payment_asset in on_initialize

* Crowdsale voucher payout + automatic payout (#796)

* Crowdsale voucher payout + automatic payout

* Fix events and build errors. Tests still broken

* Fix mock and tests

* Fix global build errors

* Add NextUnsignedAt

* Update mock + runtime

* Address PR comments

* Rename to redeem_voucher

* Fix build errors

* initialized TRN crowdsale pallet

* upd cargo.lock

* added asset_exists function to InspectExt trait

* added vec import to nft impls - fixed external dep compilation error

* updated redeem, added todos, updates structs

* Add mock + tests for initialize (#791)

* Add mock + tests for initialize

* Few tidy ups

* Add impls file, calculate_voucher_rewards, and on_initialize logic

* Update voucher_rewards calc to account for decimals partially (Needs some thinking)

* minor improvements + setup

* updated logic for burning vouchers

---------


* Pallet crowdsale upd (#793)

* added log target, removed decimals const, removed redundant enums

* fixed compilation

* fix div by zero bug

* fixed decimals burn calc logic

* updated close_sales_at with concrete example and todos

* cargo fmt --all

* Re introduce decimals, some testing work for crowdsale rounding

---------


* Crowdsale voucher rewards calculations (#794)

* added log target, removed decimals const, removed redundant enums

* fixed compilation

* fix div by zero bug

* fixed decimals burn calc logic

* updated close_sales_at with concrete example and todos

* cargo fmt --all

* Re introduce decimals, some testing work for crowdsale rounding

* Add some tests

* Add better calculation function

* Add comments and better tests

* Remove extra comment in assetsExt

* Change input type to U256 for contributions

* Added Result return type for reward calculation

---------


* Pallet crowdsale upd (#795)

* added sp io dep

* - updated SaleInformation to add vault account
- updated SaleStatus to also display block num with status variants

* - updated storage types
- updated events and errors
- added code for offchain worker
- added claim function
- refactor logic

* SaleEndBlocks back to option query, payout admin account payment_asset in on_initialize

* Crowdsale voucher payout + automatic payout (#796)

* Crowdsale voucher payout + automatic payout

* Fix events and build errors. Tests still broken

* Fix mock and tests

* Fix global build errors

* Add NextUnsignedAt

* Update mock + runtime

* Address PR comments

* Rename to redeem_voucher

* Fix build errors

* removed dependency on nft pallet for prod builds

* cargo fmt --all

* Add tests for initialize, enable and participate
Update asset to transfer not mint

* Fix build errors

* Add some tests for on_initialize

* Rename InvalidSoftCap

* Rename payment_asset and voucher

* Simplify try_mutate sale_info

* Add TODO

---------

+++ pallet/crowdsale/src/mock.rs
@@ -53,7 +53,7 @@ parameter_types! {
 	pub const MaxSalesPerBlock: u32 = 5;
 	pub const MaxConsecutiveSales: u32 = 1000;
 	pub const MaxPaymentsPerBlock: u32 = 5;
-	pub const MaxSaleDuration: u32 = 1000;
+	pub const MaxSaleDuration: u64 = 1000;
 	pub const UnsignedInterval: u32 = 10;
 }
